requirements.txt
numpy==1.24.3
Pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
wandb==0.16.3
python-dotenv==1.0.0
foolbox==3.3.4